# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Aurélien",
        "description": "C très Bo Whoulla",
        "contact": {
            "email": "admin9@fromage.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "CRUD_Book"
                ],
                "summary": "Get all books",
                "description": "Returns all books",
                "operationId": "getAllBooks",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort your résults ('title' or 'pages_nb' only)",
                        "required": false
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search your résults",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRUD_Book"
                ],
                "summary": "Create book",
                "description": "create book",
                "operationId": "createBook",
                "requestBody": {
                    "description": "Pass book information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author_id",
                                    "description",
                                    "pages_nb",
                                    "publication_year",
                                    "genre_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "string",
                                        "example": "C'est un beau titre"
                                    },
                                    "author_id": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "number",
                                        "example": "53"
                                    },
                                    "description": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "text",
                                        "example": "Description"
                                    },
                                    "pages_nb": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "number",
                                        "example": "206"
                                    },
                                    "publication_year": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "number",
                                        "example": "1999"
                                    },
                                    "genre_id": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "number",
                                        "example": "17"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "CRUD_Book"
                ],
                "summary": "Get book by ID",
                "description": "Returns book by his id",
                "operationId": "getBookByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD_Book"
                ],
                "summary": "Delete book by ID",
                "description": "Delete book by his id",
                "operationId": "deleteBookByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book was deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "CRUD_Book"
                ],
                "summary": "Update book by ID",
                "description": "Update book by his id",
                "operationId": "updateBookByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass book information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "string",
                                        "example": "C'est un beau titre update"
                                    },
                                    "author_id": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "number",
                                        "example": "54"
                                    },
                                    "description": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "text",
                                        "example": "Description DE OUF"
                                    },
                                    "pages_nb": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "number",
                                        "example": "666"
                                    },
                                    "publication_year": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "number",
                                        "example": "1990"
                                    },
                                    "genre_id": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "number",
                                        "example": "16"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}